refactor(openomni): move Manual QA drivers out of src into test/harness
Relocate the stakes, verifier-registry, completion-admission, and
existing-agent-message drivers (fixtures, scenarios, CLI mains) from
packages/openomni/src to packages/openomni/test/harness. Delete the
9-LOC verifier-registry-driver.ts shim (drivers are invoked via the
harness API or bun run on the harness path). Drop the driver exports
from the root barrel and the ledger/messaging/evidence/work-item
cluster indexes; tests import the harness directly. No open issue
Manual-QA contract names the old src paths (#215 closed; #216/#217
use apps/server/src/manual drivers). Docs updated to the new paths.
