+++ AGENTS.md
@@ -93,8 +93,8 @@ raw channel event
 | WorkItem schemas + events | `packages/protocol/src/work-item/` | `WorkItem.Info`, `Blocker`, `Evidence`, `VerificationGate`, `Status`, `deriveStatus()`, `generateHash()`, `WorkItem.Events.*`; `index.ts` is the public facade |
 | WorkItem storage interface | `packages/protocol/src/storage/index.ts` | `Storage.WorkItemSubAdapter` (get/create/compareAndSet/list/remove) |
 | WorkItemStore substrate | `packages/session/src/work-item/index.ts` | CRUD + non-completion lifecycle + blockers + evidence + dependency readiness + cycle detection; raw `complete()` is a typed refusal because product completion authority belongs in OpenOmni |
-| WorkItem completion authority | `packages/openomni/src/work-item/` | Pure durable+proposed fact fold, trusted Policy/Stakes/result/Owner authority resolver, origin projector, atomic record-before-terminal admission service, and public six-scenario Manual QA driver |
-| Windowed Stakes primitive | `packages/openomni/src/ledger/` | Deterministic consequence calculator, replay driver, criterion treatment, and per-host capability seams; WorkItem completion now consumes the Stakes resolver seam while authorized Voice remains unwired |
+| WorkItem completion authority | `packages/openomni/src/work-item/` | Pure durable+proposed fact fold, trusted Policy/Stakes/result/Owner authority resolver, origin projector, atomic record-before-terminal admission service (six-scenario Manual QA driver lives in `packages/openomni/test/harness/`) |
+| Windowed Stakes primitive | `packages/openomni/src/ledger/` | Deterministic consequence calculator, criterion treatment, and per-host capability seams (replay driver lives in `packages/openomni/test/harness/`); WorkItem completion now consumes the Stakes resolver seam while authorized Voice remains unwired |
 | Worker run records | `packages/session/src/worker-run/` | Direct DB table (worker_run_state), NOT event-sourced |
 | WorkerRun state store | `packages/session/src/worker-run/state-store.ts` | Direct DB CRUD for worker_run_state table |
 | Add LLM provider | `packages/llm/src/provider/` (`index.ts` + `sdk.ts`) + auth/transform modules as needed | Register SDK in `getSDK()`; keep provider-specific request/auth behavior out of call sites |
