+++ packages/openomni/AGENTS.md
@@ -11,9 +11,9 @@ Product kernel for OpenOmni. Builds on `@openomni/agent`, `@openomni/policy`, `@
 | `src/ingress/` | Current inbound stage: single kernel route resolution, authority facts, projection, resident/direct execution | `IngressEngine`, `resolveRoute`, `IngressEventProjector`, `IngressHandlers`, `IngressSessionResolver`, `SessionBridge`, `CronAdapter`, `resolveTarget`, `targetKey` |
 | `src/dispatch/` | Current egress/cross-boundary stage: command authorization, handler routing, PendingInteraction routing, gate-side policy stamping (#479) | `DispatchRuntime`, `DispatchRegistry`, `createDefaultDispatchRuntime` (+ `BuiltInDispatchOptions.policyResolver`), `DEFAULT_DISPATCH_MODEL` |
 | `src/policy/` | Gate-side policy resolution: actor/target labels → stamped `policyPlan` on spawn requests (production-wired by #479) | `PolicyResolver` |
-| `src/evidence/` | Read-back executors plus scoped deterministic verifier-registry and replay-conformance primitives | `ReadBackExecutor`, `VerifierRegistry`, `VerifierConformance`, `runVerifierRegistryDriver` |
-| `src/ledger/` | Deterministic windowed Stakes calculator, replay driver, criterion treatment, and host capability seams | `Stakes`, `runStakesDriver` |
-| `src/work-item/` | WorkItem completion fold, trusted authority, origin projection, atomic admission/terminal boundary, and public Manual QA driver | Public `projectCompletionOrigin`, `runCompletionAdmissionDriver`; internal `evaluateCompletion`, `createCompletionDecision`, `createCompletionAdmissionService` |
+| `src/evidence/` | Read-back executors plus scoped deterministic verifier-registry and replay-conformance primitives | `ReadBackExecutor`, `VerifierRegistry`, `VerifierConformance` |
+| `src/ledger/` | Deterministic windowed Stakes calculator, criterion treatment, and host capability seams | `Stakes` |
+| `src/work-item/` | WorkItem completion fold, trusted authority, origin projection, atomic admission/terminal boundary | Public `projectCompletionOrigin`; internal `evaluateCompletion`, `createCompletionDecision`, `createCompletionAdmissionService` |
 | `src/execution-runtime/` | Tool system, workspace, worker middleware, and scheduled job runtime | `buildWorkerMiddleware`, `WorkspaceLock`, `AgentToolProvider`, `SystemToolProvider`, `ToolProxyProvider`, `Tool`, `buildToolCatalog`, `createToolExecutor`, `defineTool`, `InjectionQueue`, `CronJobRegistry`, `CronJobRunner` |
 
 ## Architecture
@@ -83,14 +83,15 @@ Consumers should only use `@openomni/openomni` exports:
 
 - Tool system, workspace lock, worker middleware, and cron runtime from `src/execution-runtime/`
 - Scoped verifier-registry and replay-conformance library surfaces from `src/evidence/`; archived replay stays outside these primitives
-- The WorkItem completion origin projector and deterministic six-scenario Manual QA driver; the authority, fold, and terminal service stay kernel-internal
-- Windowed Stakes primitive and driver from the dedicated `@openomni/openomni/ledger` subpath; completion consumes its resolver seam while durable-ledger and Voice consumers remain separate work
+- The WorkItem completion origin projector; the authority, fold, and terminal service stay kernel-internal
+- Windowed Stakes primitive from the dedicated `@openomni/openomni/ledger` subpath; completion consumes its resolver seam while durable-ledger and Voice consumers remain separate work
 
 If a symbol is not re-exported from `src/index.ts`, treat it as private to its domain. The dormant
-`./ledger` and `./messaging` are the two subpath exceptions (`./messaging` is
-mandated by #215 Manual QA: the existing-agent message driver must be runnable
-from `@openomni/openomni/messaging`): `src/ledger/index.ts` exports `Stakes` and
-`runStakesDriver` without widening the root barrel.
+`./ledger` and `./messaging` are the two subpath exceptions: `src/ledger/index.ts`
+exports `Stakes` and `./messaging` exports the production send surface without
+widening the root barrel. The deterministic Manual QA drivers (stakes, verifier
+registry, completion admission, existing-agent messaging) live in
+`test/harness/` and are runnable with `bun run packages/openomni/test/harness/<driver>.ts`.
 
 ## Extension Points
