+++ packages/openomni/src/evidence/index.ts
@@ -1,8 +1,3 @@
 export { ReadBackExecutor } from "./read-back-executor";
 export { VerifierRegistry } from "./verifier-registry";
 export * as VerifierConformance from "./verifier-conformance";
-export { runVerifierRegistryDriver } from "./verifier-registry-driver-api";
-export type {
-  VerifierRegistryDriverExecution,
-  VerifierRegistryDriverScenario,
-} from "./verifier-registry-driver-api";
