+++ packages/openomni/src/work-item/index.ts
@@ -1,13 +1,5 @@
 import { WorkItem } from "@openomni/protocol";
 
-export {
-  CompletionAdmissionDriverScenarios,
-  runCompletionAdmissionDriver,
-} from "./completion-admission-driver.js";
-export type {
-  CompletionAdmissionDriverExecution,
-  CompletionAdmissionDriverScenario,
-} from "./completion-admission-driver.js";
 export type { WorkItemCompletionRecoveryReceipt } from "./completion-admission.js";
 
 export const CompletionSourceOrigin: typeof WorkItem.CompletionSourceOrigin =
