+++ packages/openomni/test/evidence/verifier-registry-benchmark.test.ts
@@ -1,7 +1,7 @@
 /// <reference types="bun" />
 
 import { describe, expect, test } from "bun:test";
-import { runVerifierRegistryDriver } from "../../src/evidence/verifier-registry-driver";
+import { runVerifierRegistryDriver } from "../harness/verifier-registry-driver-api";
 
 describe("verifier registry benchmark evidence", () => {
   test("derives every benchmark axis from executable fixtures", () => {
